[Sema] Error on implementation-only imported types in SPI with @_spiOnly
This used to be just a warning because it happened to work when clients
used only swiftmodules. Some projects have been using it as an
alternative to @_spiOnly imports. Now that we have an official feature
we can complete the diagnositc and error on implementation-only imports
types used in SPI.