[NFC] Extract getParameterInfo helper
The methods `importFunctionParameterList` and
`importMethodParamsAndReturnType` share a similar piece of code to
create the parameter information from the results of finding the
parameter type. In order to simplify the later refactor and to make the
changes clearer, extract the common code into a helper function and call
it from both methods, replacing the duplicated code.