[NFC] Extract getImportTypeKindForParam helper
The same piece of code was repeated in two different methods:
`importFunctionParameterList` and `importMethodParamsAndReturnType`. To
avoid repetition and simplify the later refactoring extract the code
into a helper.

The code in the two methods was happening in slightly different places,
but the modified code should still be equivalent and should not
introduce functional differences.