[Test] Help CoreGraphics_test.swift defeat the optimizer.
This test calls `CGPoint.applying` and `CGSize.applying` and ignores the return value, then checks for the call in the resulting IR. The optimizer appears to have become clever enough to notice that these calls can be eliminated entirely because their result is unused, causing the test to fail.

Instead of `let _ =`, use a `blackHole` function to make these values be used.

rdar://99489134