Handle OperandOwnerhip::NonUse in PrunedLiveness::computeSSALiveness.
Fixes a bug in the simple PrunedLiveness routine for SSA values.

Without this fix, the init_existential_ref instruction below was not
considered a lifetime-ending use because the non-use type dependednt
operand (which is not really an operand at all) would override the
forwarding operand:

  %1 = open_existential_ref %0 : $any P to $@opened("...", any P) Self
  %2 = init_existential_ref %1 : $@opened("...", any P) Self : $@opened("...", any P) Self, $AnyObject

This will be tested by silgen_cleanup_complete_ossa.sil.