refactor export-import and module-source checks to handle third-order extensions
consider this scenario:

- module A
  - struct BaseStruct
- module B
  - struct BaseStruct.InnerStruct
  - func BaseStruct.InnerStruct.doStuff()
- module C
  - `@_exported import B`

the current export-import code effectively short-circuits if the direct
parent of a symbol is from a re-exported module. however, that symbol
could itself be another extension, so the check needs to recurse
_first_, to catch this scenario. before this change,
`BaseStruct.InnerStruct.doStuff()` would be sorted into C's main module
graph, but `BaseStruct.InnerStruct` would be sorted into `C@A` instead,
creating an orphan symbol.