[CSRanking] Adjust ranking of `makeIterator` overloads in `for-in` context
Prefer an overload of `makeIterator` that satisfies
`Sequence#makeIterator` requirement when `.makeIterator()`
is implicitly injected after sequence expression in `for-in`
loop context.

This helps to avoid issues related to conditional conformances
to `Sequence` that add `makeIterator` preferred by overloading
rules that could change behavior i.e.:

```swift
extension Array where Element == Double {
  func makeIterator() -> Array<Int>.Iterator {
    return [1, 2, 3].makeIterator()
  }
}

for i in [4.0, 5.0, 6.0] {
  print(i)
}
```

`makeIterator` declared in the `Array` extension is not a witness
to `Sequence#makeIterator` and shouldn't be used because that would
change runtime behavior.

Resolves: https://github.com/apple/swift/issues/60514
Resolves: https://github.com/apple/swift/issues/60958