[move-only] Add helper method SILType::isMoveOnlyType().
This allows one to know that the type is for a first class move only type and
not a move only wrapped type.

Previously, I did this by checking type.isMoveOnly() &&
!type.isMoveOnlyWrapped(). Better to just have a helper that semantically does
this rather than using a code pattern that doesn't semantically say what it
actually does unless one knows the intent.