Factor "maybeWrapInExistential" into builders
When decoding mangled names into types, a type may be optionally
wrapped in an existential depending on the context (for example,
if a protocol appears in a function signature's requirement it's
only the protocol type, but if it appears as a type of a variable,
it's an existential). This patch centralizes the logic of whether
a type should be wrapped in an existential in a new
"maybeWrapInExistential" in the builders.