[IRGen] Emit bit for @main into entry section.
The _swift5_entry section contains a relative pointer to the entry-point
function (spelled variously _main and @main depending on context).
Support differentiating whether that entry point originates from a type
annotated @main or something else (i.e. main.swift) via a second
relative-pointer-sized field.  For now, that field is just a bit: 1 if
there is a type annotated @main, 0 otherwise.  In the future, that field
could be a relative pointer to the type descriptor for the type.