[CanonicalizeOSSALifetime] Swapped ignore meaning.
In various places, CanonicalizeOSSALifetime checks whether instructions
are "interesting enough to be worth hoisting over" for the purpose of
avoiding hoisting over instructions that would result in churning and
failing to find a fixed point when iterating in CopyPropagation.
Elsewhere, ::ignoredByDestroyHoisting is used.  Previously,
findDestroyOnCFGEdge used isIncidentalUse.  Replaced that useage with
::ignoredByDestroyHoisting.